+++ artemis/src/test/java/com/artemis/ParamArchTest.java
@@ -79,9 +79,15 @@ public class ParamArchTest {
 		public float x, y;
 	}
 	
-	public static class Size extends Component {}
-	public static class Asset extends Component {}
-	public static class HitPoints extends Component {}
+	public static class Size extends Component {
+		public float width, height;
+	}
+	public static class Asset extends Component {
+		public String path; 
+	}
+	public static class HitPoints extends Component {
+		public int hitpoints; 
+	}
 	public static class Velocity extends Component {
 		public float x, y;
 	}