+++ artemis/src/test/java/com/artemis/reference/Ship.java
@@ -1,14 +1,20 @@
 package com.artemis.reference;
 
 import com.artemis.EntityFactory;
+import com.artemis.ParamArchTest.Asset;
+import com.artemis.ParamArchTest.Position;
+import com.artemis.ParamArchTest.Size;
+import com.artemis.ParamArchTest.Velocity;
+import com.artemis.annotations.CRef;
 import com.artemis.annotations.Sticky;
 
+@CRef({Sprite.class, Cullible.class})
 public interface Ship extends EntityFactory<Ship> {
 	// method name maps Position
-	Ship position(float x, float y);
+	@CRef(Position.class) Ship position(float x, float y);
 	// parameter names must match field or setter name
-	Ship velocity(float x, float y);
-	Ship asset(String path);
-	Ship size(float width, float height);
+	@CRef(Velocity.class) Ship velocity(float x, float y);
+	@CRef(Asset.class) Ship asset(String path);
+	@CRef(Size.class) Ship size(float width, float height);
 	@Sticky Ship hitPoints(int current);
 }