+++ artemis/src/test/java/com/artemis/reference/Sprite.java
@@ -1,0 +1,7 @@
+package com.artemis.reference;
+
+import com.artemis.Component;
+
+public class Sprite extends Component {
+
+}